fixing stop service and loading dset format
